==> notebooks/1 - Data Collection.ipynb <==
# 1 - Data Collection

[How to Scrape Millions of Tweets using SNSCRAPER](https://medium.com/machine-learning-mastery/how-to-scrape-millions-of-tweets-using-snscraper-aa47cee400ec)

import snscrape.modules.twitter as sntwitter

import pandas as pd

converge = ["@converge_csu", "@experiencecnvrg"]

globe = ["@enjoyglobe", "@talk2globe"]

pldt = ["@pldthome", "@pldtenterprise", "@pldtent_cares", "@pldt_cares", "@pldt"]

tags = converge + globe + pldt

# Creating list to append tweet data to

tweets_list = []

%%time

# Using TwitterSearchScraper to scrape data and append tweets to list

for i, tweet in enumerate(sntwitter.TwitterSearchScraper(" OR ".join(tags) + " since:2022-01-01 until:2022-12-31").get_items()):
    tweets_list.append([tweet.date,
                        tweet.rawContent,
                        tweet.id,
                        tweet.user.username,
                        tweet.user.location,
                        tweet.conversationId,
                        tweet.coordinates,
                        tweet.place])

# Creating a dataframe from the tweets list above

tweets_df = pd.DataFrame(tweets_list, columns = ["date",
                                                 "rawContent",
                                                 "id",
                                                 "user_username",
                                                 "user_location",
                                                 "conversation_id",
                                                 "coordinates",
                                                 "place"])

# tweets_df.to_csv("inputs/tweets_raw.csv", index = False)
