==> notebooks/5 - Data Sampling.ipynb <==
# 5 - Data Sampling

import pandas as pd

import rt # remove tweets

df = pd.read_csv("inputs/tweets_3.csv") # import data

df.info()

df.head()

## Statistics

### Tweets by company

df.pivot_table(index = "company_x", aggfunc = "count")["content"]

round(df.pivot_table(index = "company_x", aggfunc = "count")["content"] / len(df) * 100, 2)

### Tweets by month

df.pivot_table(index = "month", aggfunc = "count")["content"]

round(df.pivot_table(index = "month", aggfunc = "count")["content"] / len(df) * 100, 2)

### Tweets by company and month

df.pivot_table(index = ["company_x", "month"], aggfunc = "count")["content"]

## Stratified Random Sampling

n = min(df.pivot_table(index = ["company_x", "month"], aggfunc = "count")["content"])

df_sample = df.groupby(["company_x", "month"]).sample(n = n, random_state = 0)

### Tweets by company and month

df_sample.pivot_table(index = ["company_x", "month"], aggfunc = "count")["content"]

rt.rt(345667, 83592)

## Export Data

# df_sample.to_csv("inputs/tweets_sample.csv", index = False)

df_sample.info()

df_sample.head()

## dataset

corpus = df_sample

corpus["partition"] = "train"

corpus = corpus[["content_10", "partition"]]

# corpus.to_csv("dataset/corpus.tsv", sep = "\t", header = False, index = False)

words = sorted(set(" ".join(corpus.content_10).split()))

with open("dataset/words.txt", mode = "w") as f:
    for word in words:
        # f.write(word + "\n")
